eventing support merged integration tests to one
